\name{eleva.a.la.potencia}
\alias{eleva.a.la.potencia}
\title{función que eleva un valor a la potencia de otro}
\description{
La función recibe dos números y eleva el primero a la potencia del segundo
}
\usage{
eleva.a.la.potencia(a, b)
}
\arguments{
  \item{a}{ número a elevar }
  \item{b}{ potencia a elevar }
}
\examples{
  z <- eleva.a.la.potencia(2, 5)
  z
}

