\name{multiplica.dos.numeros}
\alias{multiplica.dos.numeros}
\title{función que multiplica dos valores}
\description{
La función recibe dos números y los multiplica, devolviendo el resultado
}
\usage{
multiplica.dos.numeros(a, b)
}
\arguments{
  \item{a}{ un número a multiplicar }
  \item{b}{ otro número a multiplicar }
}
\examples{
  z <- multiplica.dos.numeros(2, 5)
  z
}

