\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{función que suma dos valores}
\description{
La función recibe dos números y los suma, devolviendo el resultado
}
\usage{
suma.dos.numeros(a, b)
}
\arguments{
  \item{a}{ un número a sumar }
  \item{b}{ otro número a sumar }
}
\examples{
  z <- suma.dos.numeros(2, 5)
  z
}
